+++ audit/asserts_test.go
@@ -396,6 +396,27 @@ func TestValidationAssertion(t *testing.T) {
 		t.Errorf("wrong number of errors")
 	}
 	t.Log(failures.Error())
+
+	if len(failures.Details()) != 2 {
+		t.Errorf("wrong number of details")
+	}
+	details := failures.Details()
+	fn, l, f := details[0].Location()
+	tt := details[0].Test()
+	if fn != "asserts_test.go" || l != 389 || f != "TestValidationAssertion" {
+		t.Errorf("wrong location of first detail")
+	}
+	if tt != audit.True {
+		t.Errorf("wrong test type of first detail")
+	}
+	fn, l, f = details[1].Location()
+	tt = details[1].Test()
+	if fn != "asserts_test.go" || l != 390 || f != "TestValidationAssertion" {
+		t.Errorf("wrong location of second detail")
+	}
+	if tt != audit.Equal {
+		t.Errorf("wrong test type of second detail")
+	}
 }
 
 //--------------------