+++ audit/failer.go
@@ -19,6 +19,7 @@ import (
 	"runtime"
 	"strings"
 	"sync"
+	"time"
 )
 
 //--------------------
@@ -40,11 +41,69 @@ type Failer interface {
 	Fail(test Test, obtained, expected interface{}, msgs ...string) bool
 }
 
+// FailureDetail contains detailed information of a failure.
+type FailureDetail interface {
+	// TImestamp tells when the failure has happend.
+	Timestamp() time.Time
+
+	// Locations returns file name, line number, and
+	// function name of the failure.
+	Location() (string, int, string)
+
+	// Test tells which kind of test has failed.
+	Test() Test
+
+	// Error returns the failure as error.
+	Error() error
+
+	// Message return the optional test message.
+	Message() string
+}
+
+// failureDetail implements the FailureDetail interface.
+type failureDetail struct {
+	timestamp  time.Time
+	fileName   string
+	lineNumber int
+	funcName   string
+	test       Test
+	err        error
+	message    string
+}
+
+// TImestamp implements the FailureDetail interface.
+func (d *failureDetail) Timestamp() time.Time {
+	return d.timestamp
+}
+
+// Locations implements the FailureDetail interface.
+func (d *failureDetail) Location() (string, int, string) {
+	return d.fileName, d.lineNumber, d.funcName
+}
+
+// Test implements the FailureDetail interface.
+func (d *failureDetail) Test() Test {
+	return d.test
+}
+
+// Error implements the FailureDetail interface.
+func (d *failureDetail) Error() error {
+	return d.err
+}
+
+// Message implements the FailureDetail interface.
+func (d *failureDetail) Message() string {
+	return d.message
+}
+
 // Failures collects the collected failures
 // of a validation assertion.
 type Failures interface {
 	// HasErrors returns true, if assertion failures happened.
 	HasErrors() bool
+
+	// Details returns the collected details.
+	Details() []FailureDetail
 
 	// Errors returns the so far collected errors.
 	Errors() []error
@@ -52,6 +111,10 @@ type Failures interface {
 	// Error returns the collected errors as one error.
 	Error() error
 }
+
+//--------------------
+// PANIC FAILER
+//--------------------
 
 // panicFailer reacts with a panic.
 type panicFailer struct{}
@@ -72,11 +135,22 @@ func (f panicFailer) Fail(test Test, obtained, expected interface{}, msgs ...str
 	panic(failString(test, obex, msgs...))
 }
 
+// NewPanicAssertion creates a new Assertion instance which panics if a test fails.
+func NewPanicAssertion() Assertion {
+	return NewAssertion(&panicFailer{})
+}
+
+//--------------------
+// VALIDATION FAILER
+//--------------------
+
 // validationFailer collects validation errors, e.g. when
 // validating form input data.
 type validationFailer struct {
-	mux  sync.Mutex
-	errs []error
+	mux     sync.Mutex
+	offset  int
+	details []FailureDetail
+	errs    []error
 }
 
 // HasErrors implements the Failures interface.
@@ -84,6 +158,13 @@ func (f *validationFailer) HasErrors() bool {
 	f.mux.Lock()
 	defer f.mux.Unlock()
 	return len(f.errs) > 0
+}
+
+// Details implements the Failures interface.
+func (f *validationFailer) Details() []FailureDetail {
+	f.mux.Lock()
+	defer f.mux.Unlock()
+	return f.details
 }
 
 // Errors implements the Failures interface.
@@ -106,24 +187,70 @@ func (f *validationFailer) Error() error {
 
 // IncrCallstackOffset implements the Failer interface.
 func (f *validationFailer) IncrCallstackOffset() func() {
-	return func() {}
+	f.mux.Lock()
+	defer f.mux.Unlock()
+	offset := f.offset
+	f.offset++
+	return func() {
+		f.mux.Lock()
+		defer f.mux.Unlock()
+		f.offset = offset
+	}
 }
 
 // Logf implements the Failer interface.
 func (f *validationFailer) Logf(format string, args ...interface{}) {
 	f.mux.Lock()
 	defer f.mux.Unlock()
-	backendPrintf(format+"\n", args...)
+	pc, file, line, _ := runtime.Caller(f.offset)
+	_, fileName := path.Split(file)
+	funcNameParts := strings.Split(runtime.FuncForPC(pc).Name(), ".")
+	funcNamePartsIdx := len(funcNameParts) - 1
+	funcName := funcNameParts[funcNamePartsIdx]
+	prefix := fmt.Sprintf("%s:%d %s(): ", fileName, line, funcName)
+	backendPrintf(prefix+format+"\n", args...)
 }
 
 // Fail implements the Failer interface.
 func (f *validationFailer) Fail(test Test, obtained, expected interface{}, msgs ...string) bool {
 	f.mux.Lock()
 	defer f.mux.Unlock()
+	pc, file, line, _ := runtime.Caller(f.offset)
+	_, fileName := path.Split(file)
+	funcNameParts := strings.Split(runtime.FuncForPC(pc).Name(), ".")
+	funcNamePartsIdx := len(funcNameParts) - 1
+	funcName := funcNameParts[funcNamePartsIdx]
 	obex := obexString(test, obtained, expected)
-	f.errs = append(f.errs, errors.New(failString(test, obex, msgs...)))
+	err := errors.New(failString(test, obex, msgs...))
+	detail := &failureDetail{
+		timestamp:  time.Now(),
+		fileName:   fileName,
+		lineNumber: line,
+		funcName:   funcName,
+		test:       test,
+		err:        err,
+		message:    strings.Join(msgs, " "),
+	}
+	f.details = append(f.details, detail)
+	f.errs = append(f.errs, err)
 	return false
 }
+
+// NewValidationAssertion creates a new Assertion instance which collections
+// validation failures. The returned Failures instance allows to test an access
+// them.
+func NewValidationAssertion() (Assertion, Failures) {
+	vf := &validationFailer{
+		offset:  2,
+		details: []FailureDetail{},
+		errs:    []error{},
+	}
+	return NewAssertion(vf), vf
+}
+
+//--------------------
+// TESTING FAILER
+//--------------------
 
 // Failable allows an assertion to signal a fail to an external instance
 // like testing.T or testing.B.
@@ -212,6 +339,17 @@ func (f *testingFailer) Fail(test Test, obtained, expected interface{}, msgs ...
 		f.failable.FailNow()
 	}
 	return false
+}
+
+// NewTestingAssertion creates a new Assertion instance for use with the testing
+// package. The *testing.T has to be passed as failable, the first argument.
+// shallFail controls if a failing assertion also lets fail the Go test.
+func NewTestingAssertion(f Failable, shallFail bool) Assertion {
+	return NewAssertion(&testingFailer{
+		failable:  f,
+		offset:    2,
+		shallFail: shallFail,
+	})
 }
 
 // EOF